More general interface for Metric (#200)
The method takeDerivativeWRTToTransform returned a function
Point[_2D]=>Option[DenseVector[Double]], which can be used to compute
the derivative of the metric at every point.  This exposes too much
detail of the internal working of the metric to the outside (e.g. that
the final gradient value is obtained by integrating over all the
points of the domain). This is not the case for all metrics (i.e. the
MutualInformation metric cannot be computed by integrating over
points).

We therefore changed the interface and gradient computation mechanism,
such that the metric now computes the full gradient (including the
derivative of the transformation). The consequence is, that the metric
now needs to know about the TransformationSpace, in order to compute
this full gradient.

+++ src/main/scala/scalismo/numerics/Integrator.scala
@@ -40,16 +40,16 @@ case class Integrator[D <: Dim: NDSpace](sampler: Sampler[D]) {
     val samples = sampler.sample
 
     val zeroVector = Vector.zeros[DO]
-    val sum = samples.par.map { case (pt, p) => f(pt).getOrElse(zeroVector) * (1f / p.toFloat) }.foldLeft(zeroVector)((a, b) => { a + b })
-    sum * (1f / (sampler.numberOfPoints - 1).toFloat)
+    val sum = samples.par.map { case (pt, p) => f(pt).getOrElse(zeroVector) * (1.0 / p) }.foldLeft(zeroVector)((a, b) => { a + b })
+    sum * (1f / (sampler.numberOfPoints - 1))
   }
 
-  def integrateVector(f: Function1[Point[D], Option[DenseVector[Float]]], dimensionality: Int): DenseVector[Float] = {
+  def integrateVector(f: Function1[Point[D], Option[DenseVector[Double]]], dimensionality: Int): DenseVector[Double] = {
     val samples = sampler.sample
 
-    val zeroVector = DenseVector.zeros[Float](dimensionality)
-    val sum = samples.par.map { case (pt, p) => f(pt).getOrElse(zeroVector) * (1f / p.toFloat) }.foldLeft(zeroVector)((a, b) => { a + b })
-    sum * (1f / (sampler.numberOfPoints - 1).toFloat)
+    val zeroVector = DenseVector.zeros[Double](dimensionality)
+    val sum = samples.par.map { case (pt, p) => f(pt).getOrElse(zeroVector) * (1.0 / p) }.foldLeft(zeroVector)((a, b) => { a + b })
+    sum * (1.0 / (sampler.numberOfPoints - 1))
   }
 
 }
